改gui文件结构
