+++ Aumiao-py/mainWindows.py
@@ -1,118 +1,11 @@
-"""
-1. 确保正确安装PyQt5
-首先，确保你已经正确安装了PyQt6。在命令行中运行以下命令来安装PyQt5：pip install PyQt5
-
-2. 安装Visual C++ Redistributable
-PyQt5依赖于Visual C++ Redistributable。如果没有安装，可能会导致DLL加载失败。
-你可以从Microsoft的官方网站下载(https://learn.microsoft.com/en-us/cpp/windows/latest-supported-vc-redist?view=msvc-170)
-并安装Visual C++ Redistributable。
-"""
-
 import sys  # noqa: I001
 
-# from PyQt5.QtWidgets import QApplication, QMainWindow, QPushButton, QVBoxLayout, QWidget, QLineEdit, QMessageBox
-from PyQt5.QtWidgets import *  # noqa: F403
-from src import *  # noqa: F403
+from gui import MainWindow
 
+from PyQt5.QtWidgets import QApplication
 
-class MainWindow(QMainWindow):  # noqa: F405
-	def __init__(self) -> None:
-		super().__init__()
-
-		self.setWindowTitle("毛毡工具")
-		self.setGeometry(100, 100, 400, 300)
-
-		self.layout = QVBoxLayout()  # noqa: F405
-
-		self.login_button = QPushButton("登录")  # noqa: F405
-		self.login_button.clicked.connect(self.login)
-		self.layout.addWidget(self.login_button)
-
-		self.clear_comments_button = QPushButton("清除评论")  # noqa: F405
-		self.clear_comments_button.clicked.connect(self.clear_comments)
-		self.layout.addWidget(self.clear_comments_button)
-
-		self.clear_red_point_button = QPushButton("清除邮箱红点")  # noqa: F405
-		self.clear_red_point_button.clicked.connect(self.clear_red_point)
-		self.layout.addWidget(self.clear_red_point_button)
-
-		self.reply_work_button = QPushButton("自动回复")  # noqa: F405
-		self.reply_work_button.clicked.connect(self.reply_work)
-		self.layout.addWidget(self.reply_work_button)
-
-		self.logout_button = QPushButton("登出")  # noqa: F405
-		self.logout_button.clicked.connect(self.logout)
-		self.layout.addWidget(self.logout_button)
-
-		self.central_widget = QWidget()  # noqa: F405
-		self.central_widget.setLayout(self.layout)
-		self.setCentralWidget(self.central_widget)
-
-	def login(self) -> None:
-		identity, ok1 = QInputDialog.getText(self, "登录", "请输入用户名:")  # noqa: F405
-		if ok1 and identity:
-			password, ok2 = QInputDialog.getText(self, "登录", "请输入密码:", QLineEdit.EchoMode.Password)  # noqa: F405
-			if ok2 and password:
-				try:
-					community.Login().login_token(identity=identity, password=password)  # noqa: F405
-					_data = user.Obtain().get_data_details()  # noqa: F405
-					account_data_manager = data.DataManager()  # noqa: F405
-					account_data_manager.update(
-						{
-							"ACCOUNT_DATA": {
-								"identity": identity,
-								"password": "******",
-								"id": _data["id"],
-								"nickname": _data["nickname"],
-								"create_time": _data["create_time"],
-								"description": _data["description"],
-								"author_level": _data["author_level"],
-							},
-						},
-					)
-					QMessageBox.information(self, "成功", "登录成功")  # noqa: F405
-				except Exception as e:
-					QMessageBox.critical(self, "错误", f"登录失败: {e}")  # noqa: F405
-
-	def clear_comments(self) -> None:
-		source, ok = QInputDialog.getItem(self, "清除评论", "请输入来源类型:", ["work", "post"], 0, False)  # noqa: F405, FBT003
-		if ok and source:
-			action_type, ok = QInputDialog.getItem(self, "清除评论", "请输入操作类型:", ["ads", "duplicates", "blacklist"], 0, False)  # noqa: F405, FBT003
-			if ok and action_type:
-				try:
-					client.Motion().clear_comments(source=source, action_type=action_type)  # type: ignore  # noqa: F405, PGH003
-					QMessageBox.information(self, "成功", "评论清除成功")  # noqa: F405
-				except Exception as e:
-					QMessageBox.critical(self, "错误", f"清除评论失败: {e}")  # noqa: F405
-
-	def clear_red_point(self) -> None:
-		method, ok = QInputDialog.getItem(self, "清除邮箱红点", "请输入方法:", ["nemo", "web"], 0, False)  # noqa: F405
-		if ok and method:
-			try:
-				client.Motion().clear_red_point(method=method)  # type: ignore  # noqa: F405, PGH003
-				QMessageBox.information(self, "成功", "邮箱红点清除成功")  # noqa: F405
-			except Exception as e:
-				QMessageBox.critical(self, "错误", f"清除邮箱红点失败: {e}")  # noqa: F405
-
-	def reply_work(self) -> None:
-		try:
-			client.Motion().reply_work()  # noqa: F405
-			QMessageBox.information(self, "成功", "自动回复成功")  # noqa: F405
-		except Exception as e:
-			QMessageBox.critical(self, "错误", f"自动回复失败: {e}")  # noqa: F405
-
-	def logout(self) -> None:
-		method, ok = QInputDialog.getItem(self, "登出", "请输入方法:", ["web"], 0, False)  # noqa: F405, FBT003
-		if ok and method:
-			try:
-				community.Login().logout(method=method)  # noqa: F405
-				QMessageBox.information(self, "成功", "登出成功")  # noqa: F405
-			except Exception as e:
-				QMessageBox.critical(self, "错误", f"登出失败: {e}")  # noqa: F405
-
-
-if __name__ == "__main__":
-	app = QApplication(sys.argv)  # noqa: F405
+def main() -> None:
+	app = QApplication(sys.argv)
 	window = MainWindow()
 	window.show()
 	sys.exit(app.exec())
