+++ Aumiao-py/src/__init__.py
@@ -14,7 +14,7 @@ if TYPE_CHECKING or _is_compiling:
 	from .utils import data, decorator
 
 # 修改后的 __all__ 列表,确保包含你需要的模块
-__all__ = ["client", "community", "data", "decorator", "edu", "forum", "library", "pickduck", "shop", "user", "whale", "work"]
+# __all__ = ["client", "community", "data", "decorator", "edu", "forum", "library", "pickduck", "shop", "user", "whale", "work"]
 
 # 已加载的模块缓存字典
 __loaded_modules: dict[str, ModuleType] = {}
@@ -35,13 +35,14 @@ _module_paths = {
 	"work": ".api.work", #  工作模块
 }
 
+__all__ = list(_module_paths.keys())# | __all__
 
 def __getattr__(name: str) -> ModuleType:
 	"""实现动态延迟加载"""
 	if name in __all__:
 		if name not in __loaded_modules:
 			# 动态导入模块 (实际运行时加载)
-			module = importlib.import_module(_module_paths[name], __package__) if name in _module_paths else importlib.import_module(f".{name}", __package__)
+			module = importlib.import_module(_module_paths.get(name, ".{name}"), __package__)
 			__loaded_modules[name] = module
 		return __loaded_modules[name]
 	# 如果没有找到相应的模块,抛出 AttributeError
